fix: use parseFloat instead of parse int

+++ src/controllers/finance/buyShare.ts
@@ -23,7 +23,7 @@ const buyShare: RequestHandler = async (req, res, next) => {
     return next(new AppError('amount is required', 400));
   }
 
-  const _amount = parseInt(amount);
+  const _amount = parseFloat(amount);
 
   if (!_amount || isNaN(_amount)) {
     return next(new AppError('invalid value for amount', 400));
